linux/ena: Add min kernel version to ENA_XSK_BUFF_DMA_SYNC_SINGLE_ARG
xsk_buff_dma_sync_for_cpu() was first introduced with a single argument
[1], and from kernel 5.10 [2] a second argument was added which was
removed in kernel 6.10 [3]. As a result ecc test fails because for
5.8 =< kernels < 5.10 the single arg call for xsk_buff_dma_sync_for_cpu
is successful.
However, AF XDP is only supported from kernel >= 5.10, therefore there
is no need to test this entry on kernels below it, so this commit fails
the entry test for kernels < 5.10.

[1]: https://github.com/torvalds/linux/commit/2b43470add8c8ff1e1ee28dffc5c5df97e955d09
[2]: https://github.com/torvalds/linux/commit/9647c57b11e563f5b33a49ef72b347753917c21c
[3]: https://github.com/torvalds/linux/commit/163943ac00cb31ac1a88ce5f78a7e2ead37329ec

Signed-off-by: Ofir Tabachnik <ofirt@amazon.com>