linux/ena: Remove unnecessary ifdefs from ENA_XDP_FEATURES define
Current code uses ifdefs in the definition of ENA_XDP_FEATURES
to add features only if they are supported.

This is done only for NETDEV_XDP_ACT_XSK_ZEROCOPY but as more
features will be added in the future, the ifdefs will
become more complex.

This current ifdef is in fact unnecessary.
This is because all of the currently available values in enum
netdev_xdp_act were added together in a single commit [1].
So either all of them are supported by the kernel, or none of
them is supported.

This is enforced by the ifdefs in the code that wrap all uses
of ENA_XDP_FEATURES. So if ENA_XDP_FEATURES is compiled,
all of the enums are supported.

The driver will compile in all kernels that don't support
netdev_xdp_act enums because ENA_XDP_FEATURES contents is never
compiled. The preprocessor only replaces all instances of
ENA_XDP_FEATURES with the enums, but the code is
ifdefed out before compilation begins.

This commit removes the unnecessary ifdefs.
It also removes the no-longer-necessary entries of
ENA_HAVE_NETDEV_XDP_ACT_XSK_ZEROCOPY from ECC and release script.

[1]: https://lore.kernel.org/bpf/327ad9c9868becbe1e601b580c962549c8cd81f2.1675245258.git.lorenzo@kernel.org/

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>