linux/ena: Add rx_copybreak support for XDP_PASS cases
This commit adds back the usage of the rx_copybreak mechanism, when
XDP_PASS is the verdict of the loaded XDP program.
This is due to a PPS degradation that was observed when rx_copybreak
was removed.

Current rx_copybreak handling code in ena_rx_skb() cannot be called
directly for the XDP_PASS case, because it uses the information about
the packet found in the ENA data structures to decide if it is
an rx_copybreak packet.

The packet information may not be up to date in the ENA data
structures, because the packet's head and/or tail may have changed
in the xdp program by using bpf_xdp_adjust_head/tail().
In such cases the packet head/tail are up to date only in the xdp_buff
struct.

This commit extracts the rx_copybreak handling code to its own function
and calls it with the correct xdp_buff parameters for the XDP_PASS
verdict case.

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>