linux/ena: Add support for XDP multi-buffer packets TX path
Add the code that maps the packet frags to the tx queue in
ena_xdp_tx_map_frame() similarly to what is done in ena_tx_map_skb().

To avoid code duplication between the SKB and XDP paths extract the
code that maps the frags to the tx queue into a function and use it
in both datapaths.

In case of an error, propagate the error returned by
dma_mapping_error() instead of returning a random -EINVAL.

Added 2 checks that cause dropping of packets:
1. If a packet has too many frags.
2. If the linear part of the packet is smaller than the llq header size.
   For skb such cases were handled using skb_header_pointer(). There is
   no such function for xdp at this time. So for now we just drop in
   such unlikely cases.

Signed-off-by: Arthur Kiyanovski <akiyano@amazon.com>