linux/ena: Add memory barrier on ena_xdp_xmit exit
In case a doorbell is not requested by the OS (XDP_XMIT_FLUSH), write
combine (WC) buffers might not be flushed on exit.
Later, in case other CPU refers to the same queue, it might trigger
a doorbell when not all WC buffers flushed to the device from the
first CPU, resulting in error in the device.

This patch adds a memory barrier to trigger flush of WC buffers in case
a doorbell was not requested by the OS.

Fixes: d6abd61df5ee ("linux/ena update ENA linux driver to version 2.3.0")
Signed-off-by: Evgeny Ostrovsky <evostrov@amazon.com>